Refactored SceneWidget interactors into visualization namespace
