+++ src/Open3D/GUI/VectorEdit.h
@@ -24,6 +24,8 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
+#pragma once
+
 #include "Widget.h"
 
 #include <Eigen/Geometry>
